pyproject.toml
[tool.poetry]
name = "alldaychessbot"
version = "1.0"
description = ""
authors = ["TechnoTalksDev <technotalksyt101@gmail.com>"]
license = "GNU Affero"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
motor = "^3.0.0"
colorama = "^0.4.4"
psutil = "^5.9.1"
coloredlogs = "^15.0.1"
py-cord = "2.0.0rc1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
